requirements.txt

# Web API フレームワーク
fastapi>=0.116.1
uvicorn[standard]>=0.35.0
python-multipart==0.0.6

# 環境変数
python-dotenv>=1.1.1

# LangChain パッケージ (互換性のある最新セット)
langchain>=0.2.11
langchain-community>=0.2.11
langchain-openai>=0.1.7
langgraph

# 外部サービス
openai>=1.24.0
pinecone[grpc]>=5.0.0

# データ処理
pydantic>=2.7.4
tiktoken
pypdf
pymupdf
tqdm

# Azure 関連
azure-core
azure-common
azure-ai-formrecognizer
azure-nspkg==3.0.2

# データベース
mysql-connector-python>=8.2.0
pymysql>=1.1.0 
asyncmy
aiomysql
# 認証関連
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4